+++ cmd/jiracollector/jiracollector/BUILD.bazel
@@ -11,9 +11,9 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//cmd/jiracollector/models",
+        "//pkg/jira/models",
         "//pkg/jirarepository",
         "//pkg/jiraservice",
         "//pkg/paginator",
-        "//pkg/jira/models",
     ],
 )
