--- pkg/jiracalculationsrepository/BUILD.bazel
@@ -1,12 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "jiracalculationsrepository",
-    srcs = [
-        "jiracalaculationsrepository.go",
-        "postgresjiracalaculationsrespository.go",
-    ],
-    importpath = "github.com/stebennett/squad-dashboard/pkg/jiracalculationsrepository",
-    visibility = ["//visibility:public"],
-    deps = ["//vendor/github.com/lib/pq"],
-)
